+++ epserde/examples/internal_param_zero.rs
@@ -38,7 +38,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "schema")]
     {
         let schema = unsafe { data.serialize_with_schema(&mut cursor)? };
-        println!("{}", schema.debug(cursor.as_bytes()));
+        println!("{}", schema.to_csv_with_data(cursor.as_bytes()));
         println!();
     }
     #[cfg(not(feature = "schema"))]