+++ epserde/examples/iter.rs
@@ -36,7 +36,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     {
         let i = a.clone().into_iter();
         let schema = unsafe { SerIter::from(i).serialize_with_schema(&mut cursor)? };
-        println!("{}", schema.debug(cursor.as_bytes()));
+        println!("{}", schema.to_csv_with_data(cursor.as_bytes()));
         println!();
     }
     #[cfg(not(feature = "schema"))]
@@ -95,7 +95,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "schema")]
     {
         let schema = unsafe { d.serialize_with_schema(&mut cursor)? };
-        println!("{}", schema.debug(cursor.as_bytes()));
+        println!("{}", schema.to_csv_with_data(cursor.as_bytes()));
         println!();
     }
     #[cfg(not(feature = "schema"))]