+++ epserde/examples/nested_file.rs
@@ -74,7 +74,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "schema")]
     {
         let schema = unsafe { data.serialize_with_schema(&mut file)? };
-        println!("{}", schema.debug(&std::fs::read(FILE_NAME)?));
+        println!("{}", schema.to_csv_with_data(&std::fs::read(FILE_NAME)?));
         println!();
     }
     #[cfg(not(feature = "schema"))]