+++ epserde/examples/slice.rs
@@ -34,7 +34,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "schema")]
     {
         let schema = unsafe { data.serialize_with_schema(&mut cursor)? };
-        println!("{}", schema.debug(cursor.as_bytes()));
+        println!("{}", schema.to_csv_with_data(cursor.as_bytes()));
         println!();
     }
     #[cfg(not(feature = "schema"))]
@@ -90,7 +90,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "schema")]
     {
         let schema = unsafe { data.serialize_with_schema(&mut cursor)? };
-        println!("{}", schema.debug(cursor.as_bytes()));
+        println!("{}", schema.to_csv_with_data(cursor.as_bytes()));
         println!();
     }
     #[cfg(not(feature = "schema"))]